requirements_freeze.txt
actionlib==1.14.0
angles==1.9.13
asttokens==2.4.1
attrs==24.2.0
backcall==0.2.0
base_local_planner==1.17.3
beautifulsoup4==4.12.3
bleach==6.1.0
bondpy==1.8.6
camera-calibration==1.17.0
camera-calibration-parsers==1.12.0
catkin==0.8.10
catkin-pkg==1.0.0
certifi==2024.8.30
charset-normalizer==3.4.0
controller-manager==0.20.0
controller-manager-msgs==0.20.0
cv-bridge==1.16.2
decorator==5.1.1
defusedxml==0.7.1
diagnostic-analysis==1.11.0
diagnostic-common-diagnostics==1.11.0
diagnostic-updater==1.11.0
distro==1.9.0
docopt==0.6.2
docutils==0.21.2
drone-wrapper==1.4.2
dynamic-reconfigure==1.7.3
dynamixel_sdk==3.7.51
executing==2.1.0
fastjsonschema==2.20.0
gazebo_plugins==2.9.2
gazebo_ros==2.9.2
gencpp==0.7.0
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.0
gennodejs==2.0.2
genpy==0.6.15
idna==3.10
image-geometry==1.16.2
importlib_metadata==8.5.0
interactive-markers==1.12.0
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.4
joint-state-publisher==1.15.1
joint-state-publisher-gui==1.15.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
laser_geometry==1.6.7
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mavros==1.20.0
message-filters==1.17.0
mistune==3.0.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
numpy==1.24.4
opencv-python==4.10.0.84
packaging==24.1
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pipreqs==0.5.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-qt-binding==0.4.4
PyYAML==6.0.2
pyzmq==26.2.0
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
referencing==0.35.1
requests==2.32.3
resource_retriever==1.12.7
rosbag==1.17.0
rosboost-cfg==1.15.8
rosclean==1.15.8
roscreate==1.15.8
rosgraph==1.17.0
roslaunch==1.17.0
roslib==1.15.8
roslint==0.12.0
roslz4==1.17.0
rosmake==1.15.8
rosmaster==1.17.0
rosmsg==1.17.0
rosnode==1.17.0
rosparam==1.17.0
rospkg==1.5.1
rospy==1.17.0
rosserial_python==0.9.2
rosservice==1.17.0
rostest==1.17.0
rostopic==1.17.0
rosunit==1.15.8
roswtf==1.17.0
rpds-py==0.20.0
rqt-console==0.4.12
rqt-drone-teleop==1.4.2
rqt-ground-robot-teleop==1.4.2
rqt-image-view==0.4.17
rqt-logger-level==0.4.12
rqt-moveit==0.5.11
rqt-reconfigure==0.5.5
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.15
rqt-runtime-monitor==0.5.10
rqt-rviz==0.7.0
rqt-tf-tree==0.6.4
rqt_action==0.4.9
rqt_bag==0.5.1
rqt_bag_plugins==0.5.1
rqt_dep==0.4.12
rqt_graph==0.4.14
rqt_gui==0.5.3
rqt_gui_py==0.5.3
rqt_launch==0.4.9
rqt_msg==0.4.10
rqt_nav_view==0.5.7
rqt_plot==0.4.13
rqt_pose_view==0.5.11
rqt_publisher==0.4.10
rqt_py_common==0.5.3
rqt_py_console==0.4.10
rqt_robot_steering==0.5.12
rqt_service_caller==0.4.10
rqt_shell==0.4.11
rqt_srv==0.4.9
rqt_top==0.4.10
rqt_topic==0.4.13
rqt_web==0.4.10
rviz==1.14.25
sensor-msgs==1.13.1
six==1.16.0
smach==2.5.2
smach-ros==2.5.2
smclib==1.8.6
soupsieve==2.6
stack-data==0.6.3
tf==1.13.2
tf-conversions==1.13.2
tf2-geometry-msgs==0.7.7
tf2-kdl==0.7.7
tf2-py==0.7.7
tf2-ros==0.7.7
tinycss2==1.3.0
topic-tools==1.17.0
tornado==6.4.1
traitlets==5.14.3
turtlebot3_example==1.2.5
turtlebot3_teleop==1.2.5
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
webencodings==0.5.1
xacro==1.14.18
yarg==0.1.9
zipp==3.20.2
